+++ Assets/Prefab/Game/Iris.prefab
@@ -719,6 +719,7 @@ GameObject:
   - component: {fileID: 114603672696637660}
   - component: {fileID: 114091528639082618}
   - component: {fileID: 114009191638688670}
+  - component: {fileID: 114816025624064074}
   m_Layer: 8
   m_Name: Iris
   m_TagString: Fille
@@ -2871,6 +2872,17 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   anchorPoint: {fileID: 4487648032559392}
+--- !u!114 &114816025624064074
+MonoBehaviour:
+  m_ObjectHideFlags: 1
+  m_PrefabParentObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 100100000}
+  m_GameObject: {fileID: 1486837572664584}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: e91cfeb24edff9943a477972f8376d47, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
 --- !u!114 &114983938147061660
 MonoBehaviour:
   m_ObjectHideFlags: 1
