+++ Assets/Prefab/Game/NetworkScoreManager.prefab.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: 89a1121ada339674caa7c316b6ceabf6
+timeCreated: 1521414476
+licenseType: Free
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
