+++ Assets/Scripts/NetworkScoreManager.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: e91cfeb24edff9943a477972f8376d47
-timeCreated: 1521412072
+guid: a7578bc8079bff54c9a0c077462a2a9e
+timeCreated: 1521398248
 licenseType: Free
 MonoImporter:
   externalObjects: {}
