+++ ProjectSettings/TagManager.asset
@@ -10,6 +10,8 @@ TagManager:
   - Fille
   - Doggo
   - PickableObject
+  - Car
+  - ScoreManager
   layers:
   - Default
   - TransparentFX
